\name{exmplESet}
\alias{exmplESet}
 \title{Example of Expression Set}
 \description{an ExpressionSet class contains minimum of exprs (expression matrix) of gene expression and phenoData (AnnotatedDataFrame of end point data).}
 \usage{data(exmplESet)}
<<<<<<< HEAD
 \value{an example ExpressionSet contains 105 expression features measured by U133A array for 151 subjects. The phenotype data has 8 columns for the same 151 subjects.
 }
=======
 \value{an ExpressionSet}
>>>>>>> 5b4c40047f362e5ec59f764ac963f67c77446a2d
 \keyword{misc}
