\name{exmplGeneSet}
\alias{exmplGeneSet}
 \title{Example of Gene Annotation}
 \description{An exmple of gene set collection to annotate both form of genomic data to genes. The gene names can be extracted by method of setName() and probe ids can be extracted by method of geneIds()}
 \usage{data(exmplGeneSet)}
<<<<<<< HEAD
 \value{a gene set collection of 10 genes with 319 unique U133A expression probe ids or Infinium HumanMethylation450 probe ids.}
=======
 \value{a gene set collection}
>>>>>>> 5b4c40047f362e5ec59f764ac963f67c77446a2d
 \keyword{misc}
