\name{exmplMSet}
\alias{exmplMSet}
 \title{Example of Conceptual Methylation Set}
 \description{an conceptual ExpressionSet class contains minimum of exprs (matrix) of DNA methylation and phenoData (AnnotatedDataFrame of end point data).}
 \usage{data(exmplMSet)}
 \value{an conceptual example ExpressionSet of 735 DNA methylation probe ids for 151 subjects. The phenotype data has 8 columns for the same 151 subjects}
 \keyword{misc}
