\name{exmplPat}
\alias{exmplPat}
 \title{Example of Conceptual Phenotype Pattern Definition Set}
 \description{An conceptual exmple of phenotype pattern definition set with three columns: stat.coef, stat.func, and endpt.vars; It defines an association pattern for three phenotypes.}
 \usage{data(exmplPat)}
 \value{a data frame}
 \keyword{misc}
