\name{exmplESet}
\alias{exmplESet}
 \title{Example of Expression Set}
 \description{an ExpressionSet class contains minimum of exprs (expression matrix) of gene expression and phenoData (AnnotatedDataFrame of end point data).}
 \usage{data(exmplESet)}
 \value{an example ExpressionSet contains 105 expression features measured by U133A array for 151 subjects. The phenotype data has 8 columns for the same 151 subjects.
 }
 \keyword{misc}
