\name{exmplGeneSet}
\alias{exmplGeneSet}
 \title{Example of Gene Annotation}
 \description{An exmple of gene set collection to annotate both form of genomic data to genes. The gene names can be extracted by method of setName() and probe ids can be extracted by method of geneIds()}
 \usage{data(exmplGeneSet)}
 \value{a gene set collection}
 \keyword{misc}
