\name{exmplGeneSet}
\alias{exmplGeneSet}
 \title{Example of Conceptual Gene Annotation}
 \description{An conceptual exmple of gene set collection to annotate both form of genomic data to genes. The gene names can be extracted by method of setName() and probe ids can be extracted by method of geneIds()}
 \usage{data(exmplGeneSet)}
 \value{a conceptual gene set collection of 10 genes with 319 unique U133A expression probe ids or Infinium HumanMethylation450 probe ids.}
 \keyword{misc}
