\name{exmplMSet}
\alias{exmplMSet}
 \title{Example of Methylation Set}
 \description{an ExpressionSet class contains minimum of exprs (matrix) of DNA methylation and phenoData (AnnotatedDataFrame of end point data).}
 \usage{data(exmplMSet)}
<<<<<<< HEAD
 \value{an example ExpressionSet of 735 DNA methylation probe ids for 151 subjects. The phenotype data has 8 columns for the same 151 subjects}
=======
 \value{an ExpressionSet}
>>>>>>> 5b4c40047f362e5ec59f764ac963f67c77446a2d
 \keyword{misc}
